Set default stat timestamps to the time the library was initialized.
By making the default time a constant over the execution of the program,
the stat output is still more useful than the default (i.e. the file
isn't from 1970), and behaves better with interactive editors which tend
to look for changes in the mtime to detect file changes by other programs.